rabbitmq/other_repos/01/requirements.txt
astroid==3.3.6
certifi==2024.12.14
charset-normalizer==3.4.0
dill==0.3.9
idna==3.10
isort==5.13.2
mccabe==0.7.0
pika==1.3.2
platformdirs==4.3.6
pylint==3.3.2
python-dotenv==1.0.1
requests==2.32.3
tomlkit==0.13.2
urllib3==2.2.3
